改善了id生成的问题和完成了notification的所有接口

+++ src/main/java/cn/hfut/huangshan/controller/ScenicManageController.java
@@ -45,7 +45,7 @@ public class ScenicManageController {
      * @return
      */
     @RequestMapping(value = "/ones/{adminId}",method = RequestMethod.GET)
-    public ResultObj getOnesManages(@PathVariable("adminId") int adminId){
+    public ResultObj getOnesManages(@PathVariable("adminId") long adminId){
         List<AdminScenicManage> onesManages = scenicManageService.getOnesManages(adminId);
         if (onesManages.size() > 0){
             return ResponseUtil.success(onesManages);
